feature: 业务集形态改造（平台内置转对接CMDB） #622
主机同步、Agent状态同步逻辑拆分
