+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/service/impl/sync/AppHostsUpdateHelper.java
@@ -28,15 +28,20 @@ import com.tencent.bk.job.common.redis.util.RedisKeyHeartBeatThread;
 import com.tencent.bk.job.common.util.ip.IpUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Service;
 
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * 业务主机更新辅助工具类
+ */
 @Slf4j
+@Service
 public class AppHostsUpdateHelper {
-
 
     private static final String machineIp = IpUtils.getFirstMachineIP();
     private static final String REDIS_KEY_UPDATE_APP_HOSTS_LOCK_PREFIX = "update-app-hosts-lock:";
@@ -46,6 +51,7 @@ public class AppHostsUpdateHelper {
         Collections.synchronizedMap(new HashMap<>());
     private final RedisTemplate<String, String> redisTemplate;
 
+    @Autowired
     public AppHostsUpdateHelper(RedisTemplate<String, String> redisTemplate) {
         this.redisTemplate = redisTemplate;
     }
