+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/service/impl/sync/SyncServiceImpl.java
@@ -24,31 +24,23 @@
 
 package com.tencent.bk.job.manage.service.impl.sync;
 
-import com.tencent.bk.job.common.cc.model.CcInstanceDTO;
-import com.tencent.bk.job.common.cc.sdk.CcClient;
-import com.tencent.bk.job.common.cc.sdk.CcClientFactory;
 import com.tencent.bk.job.common.constant.AppTypeEnum;
-import com.tencent.bk.job.common.constant.CcNodeTypeEnum;
 import com.tencent.bk.job.common.gse.service.QueryAgentStatusClient;
 import com.tencent.bk.job.common.model.dto.ApplicationDTO;
-import com.tencent.bk.job.common.model.dto.ApplicationHostInfoDTO;
 import com.tencent.bk.job.common.redis.util.LockUtils;
 import com.tencent.bk.job.common.redis.util.RedisKeyHeartBeatThread;
 import com.tencent.bk.job.common.util.TimeUtil;
 import com.tencent.bk.job.common.util.ip.IpUtils;
-import com.tencent.bk.job.common.util.json.JsonUtils;
 import com.tencent.bk.job.manage.config.JobManageConfig;
 import com.tencent.bk.job.manage.dao.ApplicationDAO;
 import com.tencent.bk.job.manage.dao.ApplicationHostDAO;
 import com.tencent.bk.job.manage.dao.HostTopoDAO;
 import com.tencent.bk.job.manage.manager.app.ApplicationCache;
 import com.tencent.bk.job.manage.service.ApplicationService;
-import com.tencent.bk.job.manage.service.HostService;
 import com.tencent.bk.job.manage.service.SyncService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.jooq.DSLContext;
-import org.jooq.exception.DataAccessException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.DependsOn;
@@ -85,9 +77,6 @@ public class SyncServiceImpl implements SyncService {
     private static final String REDIS_KEY_LAST_FINISH_TIME_SYNC_AGENT_STATUS = "last-finish-time-sync-agent-status";
     private static final String machineIp = IpUtils.getFirstMachineIP();
     private static final int MAX_RETRY_COUNT = 3;
-    private static final List<Long> allAppInsertFailHostIds = new ArrayList<>();
-    private static final List<Long> allAppUpdateFailHostIds = new ArrayList<>();
-    private static final List<Long> allAppDeleteFailHostIds = new ArrayList<>();
 
     static {
         List<String> keyList = Arrays.asList(REDIS_KEY_SYNC_APP_JOB_LOCK, REDIS_KEY_SYNC_HOST_JOB_LOCK,
@@ -118,7 +107,6 @@ public class SyncServiceImpl implements SyncService {
     private final String REDIS_KEY_SYNC_AGENT_STATUS_JOB_RUNNING_MACHINE = "sync-agent-status-job-running-machine";
     private final BlockingQueue<Pair<ApplicationDTO, Integer>> appHostFailQueue = new LinkedBlockingDeque<>();
     private volatile LinkedBlockingQueue<Long> extraSyncAppQueue;
-    private AppHostsUpdateHelper appHostsUpdateHelper;
     private volatile boolean enableSyncApp;
     private volatile boolean enableSyncHost;
     private volatile boolean enableSyncAgentStatus;
@@ -126,8 +114,10 @@ public class SyncServiceImpl implements SyncService {
     private HostWatchThread hostWatchThread = null;
     private HostRelationWatchThread hostRelationWatchThread = null;
     private final ApplicationCache applicationCache;
-    private final HostService hostService;
     private final BizSyncService bizSyncService;
+    private final HostSyncService hostSyncService;
+    private final AppHostsUpdateHelper appHostsUpdateHelper;
+    private final AgentStatusSyncService agentStatusSyncService;
 
     @Autowired
     public SyncServiceImpl(@Qualifier("job-manage-dsl-context") DSLContext dslContext,
@@ -139,8 +129,10 @@ public class SyncServiceImpl implements SyncService {
                            JobManageConfig jobManageConfig,
                            RedisTemplate<String, String> redisTemplate,
                            ApplicationCache applicationCache,
-                           HostService hostService,
-                           BizSyncService bizSyncService) {
+                           BizSyncService bizSyncService,
+                           HostSyncService hostSyncService,
+                           AppHostsUpdateHelper appHostsUpdateHelper,
+                           AgentStatusSyncService agentStatusSyncService) {
         this.dslContext = dslContext;
         this.applicationDAO = applicationDAO;
         this.applicationHostDAO = applicationHostDAO;
@@ -153,8 +145,10 @@ public class SyncServiceImpl implements SyncService {
         this.enableSyncHost = jobManageConfig.isEnableSyncHost();
         this.enableSyncAgentStatus = jobManageConfig.isEnableSyncAgentStatus();
         this.applicationCache = applicationCache;
-        this.hostService = hostService;
         this.bizSyncService = bizSyncService;
+        this.hostSyncService = hostSyncService;
+        this.appHostsUpdateHelper = appHostsUpdateHelper;
+        this.agentStatusSyncService = agentStatusSyncService;
         // 同步业务的线程池配置
         syncAppExecutor = new ThreadPoolExecutor(5, 5, 1L,
             TimeUnit.SECONDS, new ArrayBlockingQueue<>(20), (r, executor) ->
@@ -184,14 +178,15 @@ public class SyncServiceImpl implements SyncService {
         // 额外同步业务主机的队列与线程配置
         extraSyncAppQueue = new LinkedBlockingQueue<>(500);
         for (int i = 0; i < 3; i++) {
-            AppHostsSyncer extraAppHostsSyncer = new AppHostsSyncer(extraSyncAppQueue);
+            AppHostsSyncer extraAppHostsSyncer = new AppHostsSyncer(applicationService, hostSyncService,
+                extraSyncAppQueue);
             extraAppHostsSyncer.setName("[" + extraAppHostsSyncer.getId() + "]-extraAppHostsSyncer-" + (i + 1));
             extraAppHostsSyncer.start();
         }
-        appHostsUpdateHelper = new AppHostsUpdateHelper(redisTemplate);
         if (jobManageConfig.isEnableResourceWatch()) {
             // 开一个常驻线程监听业务资源变动事件
-            appWatchThread = new AppWatchThread(dslContext, applicationDAO, applicationService, applicationCache, redisTemplate);
+            appWatchThread = new AppWatchThread(dslContext, applicationDAO, applicationService, applicationCache,
+                redisTemplate);
             appWatchThread.start();
             // 开一个常驻线程监听主机资源变动事件
             hostWatchThread = new HostWatchThread(dslContext, applicationHostDAO, queryAgentStatusClient,
@@ -348,63 +343,8 @@ public class SyncServiceImpl implements SyncService {
     }
 
     private Future<Pair<Long, Long>> arrangeSyncAppHostsTask(ApplicationDTO applicationDTO) {
-        return syncHostExecutor.submit(() -> syncAppHostsAtOnce(applicationDTO));
-    }
-
-    private Pair<Long, Long> syncAppHostsAtOnce(ApplicationDTO applicationDTO) {
-        Long appId = applicationDTO.getId();
-        try {
-            appHostsUpdateHelper.waitAndStartAppHostsUpdating(appId);
-            return syncAppHostsIndeed(applicationDTO);
-        } catch (Throwable t) {
-            log.error("Fail to syncAppHosts of appId " + appId, t);
-            return null;
-        } finally {
-            appHostsUpdateHelper.endToUpdateAppHosts(appId);
-        }
-    }
-
-    private Pair<Long, Long> syncAppHostsIndeed(ApplicationDTO applicationDTO) {
-        Long appId = applicationDTO.getId();
-        Long cmdbInterfaceTimeConsuming = 0L;
-        Long writeToDBTimeConsuming = 0L;
-        CcClient ccClient = CcClientFactory.getCcClient();
-        StopWatch appHostsWatch = new StopWatch();
-        appHostsWatch.start("getHostsByAppInfo from CMDB");
-        Long startTime = System.currentTimeMillis();
-        log.info("begin to syncAppHosts:appId={}", appId);
-        List<ApplicationHostInfoDTO> hosts = getHostsByAppInfo(ccClient, applicationDTO);
-        cmdbInterfaceTimeConsuming += (System.currentTimeMillis() - startTime);
-        appHostsWatch.stop();
-        appHostsWatch.start("updateHosts to local DB");
-        startTime = System.currentTimeMillis();
-        refreshAppHosts(appId, hosts);
-        writeToDBTimeConsuming += (System.currentTimeMillis() - startTime);
-        appHostsWatch.stop();
-        log.info("Performance:syncAppHosts:appId={},{}", appId, appHostsWatch.toString());
-        return Pair.of(cmdbInterfaceTimeConsuming, writeToDBTimeConsuming);
-    }
-
-    private Pair<Long, Long> syncAppHostAgentStatus(Long appId) {
-        Long gseInterfaceTimeConsuming = 0L;
-        Long writeToDBTimeConsuming = 0L;
-        StopWatch appHostAgentStatusWatch = new StopWatch();
-        appHostAgentStatusWatch.start("listHostInfoByAppId");
-        List<ApplicationHostInfoDTO> localAppHosts = applicationHostDAO.listHostInfoByAppId(appId);
-        appHostAgentStatusWatch.stop();
-        appHostAgentStatusWatch.start("getAgentStatusByAppInfo from GSE");
-        Long startTime = System.currentTimeMillis();
-        hostService.fillAgentStatus(localAppHosts);
-        gseInterfaceTimeConsuming += (System.currentTimeMillis() - startTime);
-        appHostAgentStatusWatch.stop();
-        appHostAgentStatusWatch.start("updateHosts to local DB");
-        startTime = System.currentTimeMillis();
-        updateHostsInApp(appId, localAppHosts);
-        writeToDBTimeConsuming += (System.currentTimeMillis() - startTime);
-        appHostAgentStatusWatch.stop();
-        log.debug("Performance:syncAppHostAgentStatus:appId={},{}", appId,
-            appHostAgentStatusWatch.toString());
-        return Pair.of(gseInterfaceTimeConsuming, writeToDBTimeConsuming);
+        return syncHostExecutor.submit(() ->
+            hostSyncService.syncAppHostsAtOnce(applicationDTO));
     }
 
     @Override
@@ -443,15 +383,12 @@ public class SyncServiceImpl implements SyncService {
                     System.currentTimeMillis());
                 StopWatch watch = new StopWatch("syncHost");
                 watch.start("total");
-                allAppInsertFailHostIds.clear();
-                allAppUpdateFailHostIds.clear();
-                allAppDeleteFailHostIds.clear();
                 try {
                     log.info(Thread.currentThread().getName() + ":begin to sync host from cc");
                     List<ApplicationDTO> localApps = applicationDAO.listAllApps();
                     Set<Long> localAppIds =
                         localApps.stream().filter(app ->
-                            app.getAppType() == AppTypeEnum.NORMAL).map(ApplicationDTO::getId)
+                                app.getAppType() == AppTypeEnum.NORMAL).map(ApplicationDTO::getId)
                             .collect(Collectors.toSet());
                     log.info(String.format("localAppIds:%s", String.join(",",
                         localAppIds.stream().map(Object::toString).collect(Collectors.toSet()))));
@@ -482,16 +419,14 @@ public class SyncServiceImpl implements SyncService {
                             log.error("syncHost of app fail:appId=" + applicationDTO.getId(), t);
                         }
                     }
-                    log.info(Thread.currentThread().getName() + ":Finished:sync host from cc," +
+                    log.info(
+                        Thread.currentThread().getName() +
+                            ":Finished:sync host from cc," +
                             "cmdbInterfaceTimeConsuming={}ms,writeToDBTimeConsuming={}ms,rate={}",
-                        cmdbInterfaceTimeConsuming, writeToDBTimeConsuming,
-                        cmdbInterfaceTimeConsuming / (0. + writeToDBTimeConsuming));
-                    log.info(Thread.currentThread().getName() + ":Finished:Statistics:allAppInsertFailHostIds={}," +
-                            "allAppUpdateFailHostIds={},allAppDeleteFailHostIds={}", allAppInsertFailHostIds,
-                        allAppUpdateFailHostIds, allAppDeleteFailHostIds);
-                    allAppInsertFailHostIds.clear();
-                    allAppUpdateFailHostIds.clear();
-                    allAppDeleteFailHostIds.clear();
+                        cmdbInterfaceTimeConsuming,
+                        writeToDBTimeConsuming,
+                        cmdbInterfaceTimeConsuming / (0. + writeToDBTimeConsuming)
+                    );
                     // 将最后同步时间写入Redis
                     redisTemplate.opsForValue().set(REDIS_KEY_LAST_FINISH_TIME_SYNC_HOST,
                         "" + System.currentTimeMillis());
@@ -603,7 +538,8 @@ public class SyncServiceImpl implements SyncService {
     @Override
     public Boolean syncAppHosts(Long appId) {
         log.info("syncAppHosts:appId={}", appId);
-        Pair<Long, Long> pair = syncAppHostsAtOnce(applicationDAO.getAppById(appId));
+        ApplicationDTO applicationDTO = applicationDAO.getAppById(appId);
+        Pair<Long, Long> pair = hostSyncService.syncAppHostsAtOnce(applicationDTO);
         Long cmdbInterfaceTimeConsuming = pair.getFirst();
         Long writeToDBTimeConsuming = pair.getSecond();
         log.info("syncAppHosts:cmdbInterfaceTimeConsuming={},writeToDBTimeConsuming={}", cmdbInterfaceTimeConsuming,
@@ -650,32 +586,8 @@ public class SyncServiceImpl implements SyncService {
                 StopWatch watch = new StopWatch("syncAgentStatus");
                 watch.start("total");
                 try {
-                    log.info(Thread.currentThread().getName() + ":begin to sync agentStatus from GSE");
-                    List<ApplicationDTO> localApps = applicationDAO.listAllApps();
-                    Set<Long> localAppIds =
-                        localApps.stream().filter(app ->
-                            app.getAppType() == AppTypeEnum.NORMAL).map(ApplicationDTO::getId)
-                            .collect(Collectors.toSet());
-                    log.info(String.format("localAppIds:%s", String.join(",",
-                        localAppIds.stream().map(Object::toString).collect(Collectors.toSet()))));
-                    List<ApplicationDTO> localNormalApps =
-                        localApps.stream().filter(app ->
-                            app.getAppType() == AppTypeEnum.NORMAL).collect(Collectors.toList());
-                    Long gseInterfaceTimeConsuming = 0L;
-                    Long writeToDBTimeConsuming = 0L;
-                    for (ApplicationDTO applicationDTO : localNormalApps) {
-                        try {
-                            Pair<Long, Long> timeConsumingPair = syncAppHostAgentStatus(applicationDTO.getId());
-                            gseInterfaceTimeConsuming += timeConsumingPair.getFirst();
-                            writeToDBTimeConsuming += timeConsumingPair.getSecond();
-                        } catch (Throwable t) {
-                            log.error("syncAgentStatus of app fail:appId=" + applicationDTO.getId(), t);
-                        }
-                    }
-                    log.info(Thread.currentThread().getName() + ":Finished:sync agentStatus from GSE," +
-                            "gseInterfaceTimeConsuming={}ms,writeToDBTimeConsuming={}ms,rate={}",
-                        gseInterfaceTimeConsuming, writeToDBTimeConsuming,
-                        gseInterfaceTimeConsuming / (0. + writeToDBTimeConsuming));
+                    // 从GSE同步Agent状态
+                    agentStatusSyncService.syncAgentStatusFromGSE();
                     // 将最后同步时间写入Redis
                     redisTemplate.opsForValue().set(REDIS_KEY_LAST_FINISH_TIME_SYNC_AGENT_STATUS,
                         "" + System.currentTimeMillis());
@@ -692,299 +604,6 @@ public class SyncServiceImpl implements SyncService {
             LockUtils.releaseDistributedLock(REDIS_KEY_SYNC_AGENT_STATUS_JOB_LOCK, machineIp);
         }
         return 1L;
-    }
-
-    private List<ApplicationHostInfoDTO> getHostsByAppInfo(CcClient ccClient, ApplicationDTO applicationDTO) {
-        List<CcInstanceDTO> ccInstanceDTOList = new ArrayList<>();
-        ccInstanceDTOList.add(new CcInstanceDTO(CcNodeTypeEnum.APP.getType(), applicationDTO.getId()));
-        List<ApplicationHostInfoDTO> applicationHostInfoDTOList = ccClient.getHosts(applicationDTO.getId(),
-                                                                                    ccInstanceDTOList);
-        // 获取Agent状态
-        hostService.fillAgentStatus(applicationHostInfoDTOList);
-        return applicationHostInfoDTOList;
-    }
-
-    private void updateHostsInApp(Long appId, List<ApplicationHostInfoDTO> updateList) {
-        StopWatch watch = new StopWatch();
-        watch.start("updateAppHostInfo");
-        // 更新主机
-        long updateCount = 0L;
-        List<Long> updateHostIds = new ArrayList<>();
-        long errorCount = 0L;
-        List<Long> errorHostIds = new ArrayList<>();
-        long notChangeCount = 0L;
-        boolean batchUpdated = false;
-        try {
-            // 尝试批量更新
-            if (!updateList.isEmpty()) {
-                applicationHostDAO.batchUpdateAppHostInfoByHostId(dslContext, updateList);
-            }
-            batchUpdated = true;
-        } catch (Throwable throwable) {
-            if (throwable instanceof DataAccessException) {
-                String errorMessage = throwable.getMessage();
-                if (errorMessage.contains("Duplicate entry") && errorMessage.contains("PRIMARY")) {
-                    log.info("Fail to batchUpdateAppHostInfoByHostId, try to update one by one");
-                } else {
-                    log.warn("Fail to batchUpdateAppHostInfoByHostId, try to update one by one.", throwable);
-                }
-            } else {
-                log.warn("Fail to batchUpdateAppHostInfoByHostId, try to update one by one..", throwable);
-            }
-            // 批量更新失败，尝试逐条更新
-            for (ApplicationHostInfoDTO hostInfoDTO : updateList) {
-                try {
-                    if (!applicationHostDAO.existAppHostInfoByHostId(dslContext, hostInfoDTO)) {
-                        applicationHostDAO.updateAppHostInfoByHostId(dslContext, hostInfoDTO.getAppId(), hostInfoDTO);
-                        updateCount += 1;
-                        updateHostIds.add(hostInfoDTO.getHostId());
-                    } else {
-                        notChangeCount += 1;
-                    }
-                } catch (Throwable t) {
-                    log.error(String.format("updateHost fail:appId=%d,hostInfo=%s", appId, hostInfoDTO), t);
-                    errorCount += 1;
-                    errorHostIds.add(hostInfoDTO.getHostId());
-                }
-            }
-        }
-        watch.stop();
-        if (!batchUpdated) {
-            watch.start("log updateAppHostInfo");
-            allAppUpdateFailHostIds.addAll(errorHostIds);
-            log.info("Update host of appId={},errorCount={},updateCount={},notChangeCount={},errorHostIds={}," +
-                    "updateHostIds={}", appId, errorCount, updateCount, notChangeCount,
-                errorHostIds
-                , updateHostIds);
-            watch.stop();
-        }
-        log.debug("Performance:updateHostsInApp:appId={},{}", appId, watch.prettyPrint());
-    }
-
-    private void deleteHostsFromApp(Long appId, List<ApplicationHostInfoDTO> deleteList) {
-        StopWatch watch = new StopWatch();
-        // 删除主机
-        watch.start("deleteAppHostInfo");
-        List<Long> deleteFailHostIds = new ArrayList<>();
-        boolean batchDeleted = false;
-        try {
-            // 尝试批量删除
-            if (!deleteList.isEmpty()) {
-                applicationHostDAO.batchDeleteAppHostInfoById(dslContext, appId,
-                    deleteList.stream().map(ApplicationHostInfoDTO::getHostId).collect(Collectors.toList()));
-            }
-            batchDeleted = true;
-        } catch (Throwable throwable) {
-            log.warn("Fail to batchDeleteAppHostInfoById, try to delete one by one", throwable);
-            // 批量删除失败，尝试逐条删除
-            for (ApplicationHostInfoDTO applicationHostInfoDTO : deleteList) {
-                try {
-                    applicationHostDAO.deleteAppHostInfoById(dslContext, appId, applicationHostInfoDTO.getHostId());
-                } catch (Throwable t) {
-                    log.error("deleteHost fail:appId={},hostInfo={}", appId,
-                        applicationHostInfoDTO, t);
-                    deleteFailHostIds.add(applicationHostInfoDTO.getHostId());
-                }
-            }
-        }
-        watch.stop();
-        if (!batchDeleted) {
-            watch.start("log deleteAppHostInfo");
-            if (!deleteFailHostIds.isEmpty()) {
-                allAppDeleteFailHostIds.addAll(deleteFailHostIds);
-                log.warn(String.format("appId=%s,deleteFailHostIds.size=%d,deleteFailHostIds=%s",
-                    appId, deleteFailHostIds.size(), String.join(",",
-                        deleteFailHostIds.stream().map(Object::toString).collect(Collectors.toSet()))));
-            }
-            watch.stop();
-        }
-        log.debug("Performance:deleteHostsFromApp:appId={},{}", appId, watch.prettyPrint());
-    }
-
-    private boolean insertOrUpdateOneAppHost(Long appId, ApplicationHostInfoDTO infoDTO) {
-        try {
-            applicationHostDAO.insertAppHostInfo(dslContext, infoDTO);
-        } catch (DataAccessException e) {
-            String errorMessage = e.getMessage();
-            if (errorMessage.contains("Duplicate entry") && errorMessage.contains("PRIMARY")) {
-                log.warn(String.format(
-                    "insertHost fail, try to update:Duplicate entry:appId=%d," +
-                        "insert hostInfo=%s, old " +
-                        "hostInfo=%s", appId, infoDTO,
-                    applicationHostDAO.getHostById(infoDTO.getHostId())), e);
-                try {
-                    // 插入失败了就应当更新，以后来的数据为准
-                    applicationHostDAO.updateAppHostInfoByHostId(dslContext, appId, infoDTO);
-                } catch (Throwable t) {
-                    log.error(String.format("update after insert fail:appId=%d,hostInfo=%s", appId, infoDTO), t);
-                    return false;
-                }
-            } else {
-                log.error(String.format("insertHost fail:appId=%d,hostInfo=%s", appId, infoDTO), e);
-                return false;
-            }
-        } catch (Throwable t) {
-            log.error(String.format("insertHost fail:appId=%d,hostInfo=%s", appId, infoDTO), t);
-            return false;
-        }
-        return true;
-    }
-
-    private void insertHostsToApp(Long appId, List<ApplicationHostInfoDTO> insertList) {
-        StopWatch watch = new StopWatch();
-        // 插入主机
-        watch.start("insertAppHostInfo");
-        List<Long> insertFailHostIds = new ArrayList<>();
-        boolean batchInserted = false;
-        try {
-            //尝试批量插入
-            if (!insertList.isEmpty()) {
-                applicationHostDAO.batchInsertAppHostInfo(dslContext, insertList);
-            }
-            batchInserted = true;
-        } catch (Throwable throwable) {
-            if (throwable instanceof DataAccessException) {
-                String errorMessage = throwable.getMessage();
-                if (errorMessage.contains("Duplicate entry") && errorMessage.contains("PRIMARY")) {
-                    log.info("Fail to batchInsertAppHostInfo, try to insert one by one");
-                } else {
-                    log.warn("Fail to batchInsertAppHostInfo, try to insert one by one.", throwable);
-                }
-            } else {
-                log.warn("Fail to batchInsertAppHostInfo, try to insert one by one..", throwable);
-            }
-            //批量插入失败，尝试逐条插入
-            for (ApplicationHostInfoDTO infoDTO : insertList) {
-                if (!insertOrUpdateOneAppHost(appId, infoDTO)) {
-                    insertFailHostIds.add(infoDTO.getHostId());
-                }
-            }
-        }
-        watch.stop();
-        if (!batchInserted) {
-            watch.start("log insertAppHostInfo");
-            if (!insertFailHostIds.isEmpty()) {
-                allAppInsertFailHostIds.addAll(insertFailHostIds);
-                log.warn(String.format("appId=%s,insertFailHostIds.size=%d,insertFailHostIds=%s",
-                    appId, insertFailHostIds.size(), String.join(",",
-                        insertFailHostIds.stream().map(Object::toString).collect(Collectors.toSet()))));
-            }
-            watch.stop();
-        }
-        log.debug("Performance:insertHostsToApp:appId={},{}", appId, watch.prettyPrint());
-    }
-
-    private List<ApplicationHostInfoDTO> computeInsertList(
-        Long appId,
-        Set<Long> localAppHostIds,
-        List<ApplicationHostInfoDTO> applicationHostInfoDTOList
-    ) {
-        StopWatch watch = new StopWatch();
-        List<ApplicationHostInfoDTO> insertList =
-            applicationHostInfoDTOList.stream().filter(applicationHostInfoDTO ->
-                !localAppHostIds.contains(applicationHostInfoDTO.getHostId())).collect(Collectors.toList());
-        watch.start("log insertList");
-        log.info(String.format("appId=%s,insertHostIds=%s", appId, String.join(",",
-            insertList.stream().map(ApplicationHostInfoDTO::getHostId).map(Object::toString)
-                .collect(Collectors.toSet()))));
-        watch.stop();
-        if (watch.getTotalTimeMillis() > 1000) {
-            log.warn("Write log too slow, {}", watch.prettyPrint());
-        }
-        return insertList;
-    }
-
-    private List<ApplicationHostInfoDTO> computeUpdateList(
-        Long appId,
-        Set<Long> localAppHostIds,
-        List<ApplicationHostInfoDTO> applicationHostInfoDTOList
-    ) {
-        StopWatch watch = new StopWatch();
-        List<ApplicationHostInfoDTO> updateList =
-            applicationHostInfoDTOList.stream().filter(applicationHostInfoDTO ->
-                localAppHostIds.contains(applicationHostInfoDTO.getHostId())).collect(Collectors.toList());
-        watch.start("log updateList");
-        log.info(String.format("appId=%s,updateHostIds=%s", appId, String.join(",",
-            updateList.stream().map(ApplicationHostInfoDTO::getHostId)
-                .map(Object::toString).collect(Collectors.toSet()))));
-        watch.stop();
-        if (watch.getTotalTimeMillis() > 1000) {
-            log.warn("Write log too slow, {}", watch.prettyPrint());
-        }
-        return updateList;
-    }
-
-    private List<ApplicationHostInfoDTO> computeDeleteList(
-        Long appId,
-        Set<Long> ccAppHostIds,
-        List<ApplicationHostInfoDTO> localAppHosts
-    ) {
-        StopWatch watch = new StopWatch();
-        List<ApplicationHostInfoDTO> deleteList =
-            localAppHosts.stream().filter(applicationHostInfoDTO ->
-                !ccAppHostIds.contains(applicationHostInfoDTO.getHostId())).collect(Collectors.toList());
-        watch.start("log deleteList");
-        log.info(String.format("appId=%s,deleteHostIds=%s", appId, String.join(",",
-            deleteList.stream().map(ApplicationHostInfoDTO::getHostId).map(Object::toString)
-                .collect(Collectors.toSet()))));
-        watch.stop();
-        if (watch.getTotalTimeMillis() > 1000) {
-            log.warn("Write log too slow, {}", watch.prettyPrint());
-        }
-        return deleteList;
-    }
-
-    private int refreshAppHosts(Long appId,
-                                List<ApplicationHostInfoDTO> applicationHostInfoDTOList) {
-        StopWatch watch = new StopWatch();
-        //找出要删除的/更新的/新增的分别处理
-        //对比库中数据与接口数据
-        watch.start("listHostInfoByAppId");
-        List<ApplicationHostInfoDTO> localAppHosts = applicationHostDAO.listHostInfoByAppId(appId);
-        watch.stop();
-        watch.start("mapTo ccAppHostIds");
-        Set<Long> ccAppHostIds =
-            applicationHostInfoDTOList.stream().map(ApplicationHostInfoDTO::getHostId).collect(Collectors.toSet());
-        watch.stop();
-        watch.start("mapTo localAppHostIds");
-        Set<Long> localAppHostIds =
-            localAppHosts.stream().map(ApplicationHostInfoDTO::getHostId).collect(Collectors.toSet());
-        watch.stop();
-        watch.start("log ccAppHostIds");
-        log.info(String.format("appId=%s,ccAppHostIds=%s", appId, String.join(",",
-            ccAppHostIds.stream().map(Object::toString).collect(Collectors.toSet()))));
-        watch.stop();
-        watch.start("log localAppHostIds");
-        log.info(String.format("appId=%s,localAppHostIds=%s", appId, String.join(",",
-            localAppHostIds.stream().map(Object::toString).collect(Collectors.toSet()))));
-        watch.stop();
-        watch.start("compute insertList");
-        List<ApplicationHostInfoDTO> insertList = computeInsertList(appId, localAppHostIds, applicationHostInfoDTOList);
-        watch.stop();
-        watch.start("compute updateList");
-        List<ApplicationHostInfoDTO> updateList = computeUpdateList(appId, localAppHostIds, applicationHostInfoDTOList);
-        watch.stop();
-        watch.start("compute deleteList");
-        List<ApplicationHostInfoDTO> deleteList = computeDeleteList(appId, ccAppHostIds, localAppHosts);
-        watch.stop();
-        watch.start("deleteHostsFromApp");
-        // 需要删除的主机
-        deleteHostsFromApp(appId, deleteList);
-        watch.stop();
-        watch.start("insertHostsToApp");
-        // 需要新增的主机
-        insertHostsToApp(appId, insertList);
-        watch.stop();
-        watch.start("updateHostsInApp");
-        // 需要更新的主机
-        updateHostsInApp(appId, updateList);
-        watch.stop();
-        if (watch.getTotalTimeMillis() > 10000) {
-            log.info("Performance:refreshAppHosts:appId={},{}", appId, watch.prettyPrint());
-        } else {
-            log.debug("Performance:refreshAppHosts:appId={},{}", appId, watch.prettyPrint());
-        }
-        return 1;
     }
 
     private Long getLastFinishTime(String key) {
@@ -1011,31 +630,4 @@ public class SyncServiceImpl implements SyncService {
         return getLastFinishTime(REDIS_KEY_LAST_FINISH_TIME_SYNC_AGENT_STATUS);
     }
 
-    class AppHostsSyncer extends Thread {
-
-        volatile BlockingQueue<Long> queue;
-
-        public AppHostsSyncer(BlockingQueue<Long> queue) {
-            this.queue = queue;
-        }
-
-        @Override
-        public void run() {
-            while (true) {
-                try {
-                    Long appId = queue.take();
-                    Pair<Long, Long> timeConsumingPair = syncAppHostsAtOnce(applicationService.getAppByAppId(appId));
-                    Long cmdbInterfaceTimeConsuming = timeConsumingPair.getFirst();
-                    Long writeToDBTimeConsuming = timeConsumingPair.getSecond();
-                    log.info("Sync appHosts of {}:cmdbInterfaceTimeConsuming={}ms,writeToDBTimeConsuming={}ms," +
-                            "rate={}", appId, cmdbInterfaceTimeConsuming, writeToDBTimeConsuming,
-                        cmdbInterfaceTimeConsuming / (0. + writeToDBTimeConsuming));
-                } catch (InterruptedException e) {
-                    log.warn("queue.take interrupted", e);
-                } catch (Throwable t) {
-                    log.warn("Fail to syncAppHostsAtOnce", t);
-                }
-            }
-        }
-    }
 }
